Añadir pantalla de listado de juegos

+++ src/main/java/samuraisword/game/GameController.java
@@ -1,5 +1,6 @@
 package samuraisword.game;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -77,6 +78,7 @@ public class GameController {
 		game.setGamePhase(GamePhase.LOBBY);
 		UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 		User user = userService.findUser(userDetails.getUsername()).get();
+		game.setCreator(user);
 		Player player = new Player();
 		player.setUser(user);
 		player.setGame(game);
@@ -138,35 +140,39 @@ public class GameController {
 	public String initGame(@PathVariable("id_game") int gameId, Map<String, Object> model) {
 		String view = "redirect:/game/continue/"+gameId;
 		Game game = GameSingleton.getInstance().getMapGames().get(gameId);
-		game.setGamePhase(GamePhase.MAIN);
-		invitationService.deleteInvitationsByGame(game);
-		game.setDeck(cardService.createDeck());
-		game.setDiscardPile(new ArrayList<>());
+		if(game.getGamePhase().equals(GamePhase.LOBBY)) {
+			game.setStartDate(new Date());
+			invitationService.deleteInvitationsByGame(game);
+			game.setDeck(cardService.createDeck());
+			game.setDiscardPile(new ArrayList<>());
 
-		List<Player> players = game.getListPlayers();
+			List<Player> players = game.getListPlayers();
 
-		gameService.asignCharacterAndHearts(players);
-		gameService.asignRolAndHonor(players);
-		gameService.asignOrder(players);
+			gameService.asignCharacterAndHearts(players);
+			gameService.asignRolAndHonor(players);
+			gameService.asignOrder(players);
+			
+
+			game.setListPlayers(players);
+			game.setCurrentPlayer(players.get(0));
+			game.getCurrentPlayer().setWeaponBonus(1);
 		
 
-		game.setListPlayers(players);
-		game.setCurrentPlayer(players.get(0));
-		game.getCurrentPlayer().setWeaponBonus(1);
-	
-
-		for (Player player : game.getListPlayers()) {
-			player.setGame(game);
-			player.setEquipment(new ArrayList<>());
-			playerService.savePlayer(player);
-			if(player.getCharacter().getName().equals("Goemon")) {
-				player.setWeaponBonus(2);
+			for (Player player : game.getListPlayers()) {
+				player.setGame(game);
+				player.setEquipment(new ArrayList<>());
+				playerService.savePlayer(player);
+				if(player.getCharacter().getName().equals("Goemon")) {
+					player.setWeaponBonus(2);
+				}
 			}
-		}
 
-		gameService.asignCards(game.getDeck(), players);
-		gameService.processDrawPhase(game);
-		game.setGamePhase(GamePhase.MAIN);
+			gameService.asignCards(game.getDeck(), players);
+			gameService.processDrawPhase(game);
+			game.setGamePhase(GamePhase.MAIN);
+			gameService.saveGame(game);
+		}
+		
 		return view;
 	}
 
@@ -218,6 +224,7 @@ public class GameController {
 				view = endGame(game, model);
 				Rol winnerRol = gameService.calcWinners(game);
 				game.setWonPlayers(new ArrayList<User>());
+				game.setEndDate(new Date());
 				for(Player p: game.getListPlayers()) {
 					if(p.getRol().equals(winnerRol) || (winnerRol.equals(Rol.SAMURAI) && p.getRol().equals(Rol.SHOGUN))) {
 					game.getWonPlayers().add(p.getUser());
@@ -232,6 +239,7 @@ public class GameController {
 	public String endGame(Game game, Map<String, Object> model) {
 		Rol winnerRol = gameService.calcWinners(game);
 		model.put("winnerRol", winnerRol);
+		game.setEndDate(new Date());
 		return "/game/endgame";
 	}
 
@@ -245,6 +253,7 @@ public class GameController {
 			Game game = GameSingleton.getInstance().getMapGames().get(gameId);
 			if(!gameService.checkAllPlayersHavePositiveHonor(game)) {
 				view = endGame(game, model);
+				game.setEndDate(new Date());
 				Rol winnerRol = gameService.calcWinners(game);
 				game.setWonPlayers(new ArrayList<User>());
 				for(Player p: game.getListPlayers()) {
@@ -274,11 +283,24 @@ public class GameController {
 			}
 			
 			model.put("game", game);
+			Boolean b=false;
+			for(Player p:game.getListPlayers()) {
+				if(p.getUser().getUsername().equals(user.getUsername())) {
+					b=true;
+				}
+			}
+			model.put("b", b);
+			if(b) {
+				Player POVplayer = game.getListPlayers().stream()
+						.filter(p -> p.getUser().getUsername().equals(user.getUsername()))
+						.findFirst().get();
+				model.put("POVplayer", POVplayer);
+			}else {
+				Player POVplayer = new Player();
+				POVplayer.setUser(user);
+				model.put("POVplayer", POVplayer);
+			}
 			
-			Player POVplayer = game.getListPlayers().stream()
-					.filter(p -> p.getUser().getUsername().equals(user.getUsername()))
-					.findFirst().get();
-			model.put("POVplayer", POVplayer);
 			model.put("gameStatus", game.getGamePhase().toString());
 			String discardImage;
 			if(game.getDiscardPile().isEmpty()) {
